// src/pages/Operations/components/OperationsReport/constants/urlSearchParams.ts
export enum OperationsReportUrlSearchParams {
  startDate = 'startDate',
  endDate = 'endDate',
  operationType = 'operationType',
  operationKind = 'operationKind',
  contractorId = 'contractorId',
  contractorName = 'contractorName',
  articleId = 'articleId',
  articleName = 'articleName',
  minAmount = 'minAmount',
  maxAmount = 'maxAmount',
  search = 'search',
}